+++ CUSTOM_FILES/save_for_now/b4_comletely_adding_new_object_modules_module_configurator.rb
@@ -91,7 +91,7 @@ module Marskal
           yield(configuration)
         end
 
-        def config_override(p_options= {})
+        def  mcfg_config_override(p_options= {})
           configuration.instance_values.merge(p_options).symbolize_keys
         end