+++ lib/examples/defaults_examples.rb
@@ -1,3 +1,6 @@
+require 'pp'
+
+
 module DefaultsExperiments
   include Marskal::ModuleConfigurator       #this will provide access to the ModuleConfigurator methods
 
@@ -26,9 +29,9 @@ module DefaultsExperiments
 
   def self.huh
     # Now lets add some custom dynamic settings
-    setup({ color: 'RED', material: 'Steel', sides: 20 }, set_defaults: true )
-    reset
-    c = @configuration
+    mcfg_setup({ color: 'RED', material: 'Steel', sides: 20 }, set_defaults: true )
+     mcfg_reset
+    c =  @mcfg_config
     c
   end
 
@@ -40,22 +43,22 @@ module DefaultsExperiments
     )
 
     # Now lets add some custom dynamic settings
-    setup({ color: 'RED', material: 'Steel'}, set_defaults: true )
+    mcfg_setup({ color: 'RED', material: 'Steel'}, set_defaults: true )
 
     #now lets change one of our custom settings defaults
     mcfg_set_defaults(material: 'Plastic')
 
     # Now lets add some a dynamic settings and specifically ask NOT to set defaults (set_defaults: false)
-    setup({ weight: 'Heavy' }, set_defaults: false )
+    mcfg_setup({ weight: 'Heavy' }, set_defaults: false )
 
     # Now lets add some a dynamic settings and allow the default to be used (default: set_defaults = false)
-    setup({ bet: '$100' })
+    mcfg_setup({ bet: '$100' })
 
     pp(mcfg_defaults) #prints {:sides=>10, :players=>5, :color=>"RED", :material=>"Plastic"}
 
-    reset(remove_added_attributes: false)  #now lets reset, but keep our attributes and defaults
+     mcfg_reset(remove_added_attributes: false)  #now lets reset, but keep our attributes and defaults
 
-    pp(@configuration)  #the output below is produced
+    pp( @mcfg_config)  #the output below is produced
     # #<DefaultsExperiments::Configuration:0x000000045f7838
     #     @color="RED",
     #     @material="Plastic",
@@ -67,8 +70,8 @@ module DefaultsExperiments
 
     #Notice How 'weight' and 'bet' do not appear. HOWEVER, they are still accessible
     # Let's test that out
-    @configuration.bet = '$200'
-    pp(@configuration) #this produces the output below
+     @mcfg_config.bet = '$200'
+    pp( @mcfg_config) #this produces the output below
     # #<DefaultsExperiments::Configuration:0x0000000437f600
     #     @bet="$200",
     #     @color="RED",
@@ -81,9 +84,9 @@ module DefaultsExperiments
 
     #Now lets do a default reset which will wipe out access to all dynamically added fields
     #But will leave the predefined class values and apply any defaults to them
-    reset #reset using defaults
+     mcfg_reset #reset using defaults
 
-    pp(@configuration) #this produces the output below
+    pp( @mcfg_config) #this produces the output below
     # #<DefaultsExperiments::Configuration:0x000000044a6628
     #     @num_dice=1,
     #     @players=5,
@@ -92,15 +95,15 @@ module DefaultsExperiments
     #     @target_score=25>
 
     begin
-      @configuration.weight = 1
+       @mcfg_config.weight = 1
     rescue NameError => error
       puts "Error: After Reset, the dynamically added variable is no longer available"
     end
 
     #now lets reset to original. We will not apply any of the established_defaults
-    reset(apply_defaults: false)
+     mcfg_reset(apply_defaults: false)
 
-    pp(@configuration) #this produces the output below
+    pp( @mcfg_config) #this produces the output below
     # #<DefaultsExperiments::Configuration:0x00000004494c48
     #     @num_dice=1,
     #     @players=2,
@@ -110,7 +113,7 @@ module DefaultsExperiments
 
 
 
-    @configuration
+     @mcfg_config
   end
 
 end